Revert "Avoid unnecessary blank lines in output"
This reverts commit 22dc9bad

This resulted in even weirder output to the console. The ProgressHelper
has a weird relationship with the content that comes immediately after
it, and it sometimes forces the next content to follow a blank line or
to be on the same line but shoved far to the right. It might be best to
just live with extra blank lines for the moment.