Drop 'automatically clear lock' threshold to 2 hours
Since the datetime on the lock file is updated at every endpoint rather
than just when the update_stats process starts, it's unreasonable for
a single endpoint to be anticipated to be processed for 24 hours. 2
hours accommodates edge-cases with endpoints with many data points and
allows the every-three-hours cron jobs that run update_stats to more
quickly resume incomplete imports.