Remove "Price of Gold" endpoint group
I’ve discovered that the “price of gold” data is, unfortunately and inexplicably, not available through the FRED API and can only be viewed by visiting their website. Attempts to retrieve that data from the API result in the message “Not Found. The data for this series is not available via the API; the data can be found on the FRED website.” I have removed that page from the website. If FRED allows this data to be accessed through their API in the future, this page can be re-added.