Also use realtime_end to determine if an update is available
In both API results and the API documentation (https://fred.stlouisfed.org/docs/api/fred/series.html),
FRED sometimes returns results with last_updated an earlier date than
realtime_end (which is the latest date associated with available
statistics). I don't have an explanation for that, but relying on both
dates to determine if an update is available is an acceptable
workaround.